Commit core flow update!!!

+++ source/FootballReservationSystem/src/main/java/com/services/TimeSlotServices.java
@@ -1,15 +1,15 @@
 package com.services;
 
 import com.dto.InputReservationDTO;
+import com.dto.TimeSlotDTO;
 import com.entity.*;
 import com.repository.TimeSlotRepository;
 import com.utils.DateTimeUtils;
+import org.hibernate.validator.internal.xml.FieldType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
+import java.util.*;
 
 /**
  * Created by MinhQuy on 9/29/2017.
@@ -34,7 +34,6 @@ public class TimeSlotServices {
     @Autowired
     FriendlyMatchServices friendlyMatchServices;
 
-
     public List<TimeSlotEntity> findUpcomingReservationByDate(String dateString, int fieldOwnerId, int fieldTypeId) {
         Date targetDate = DateTimeUtils.convertFromStringToDate(dateString);
         AccountEntity accountEntity = accountServices.findAccountEntityById(fieldOwnerId, "owner");
@@ -56,22 +55,34 @@ public class TimeSlotServices {
             for (AccountEntity fieldOwnerEntity : fieldOwnerList) {
                 // đếm chủ sân có bao nhiêu sân loại đó
                 int numberOfField = fieldServices.countNumberOfFieldByFieldOwnerAndFieldType(fieldOwnerEntity, fieldTypeEntity);
-                // get list timeEnable theo chủ sân, loại sân và theo ngày
+                // get list timeEnable theo chủ sân, loại sân và theo ngày order by theo thời gian từ nhỏ đến lớn
                 List<TimeEnableEntity> timeEnableEntityList = timeEnableServices.findTimeEnableByFieldOwnerTypeAndDate(fieldOwnerEntity, fieldTypeEntity, dayInWeek);
-
-                for (TimeEnableEntity timeEnableEntity : timeEnableEntityList) {
+                if (timeEnableEntityList.size() != 0) {
                     TimeSlotEntity timeSlotEntity = new TimeSlotEntity();
                     timeSlotEntity.setDate(date);
                     timeSlotEntity.setFieldOwnerId(fieldOwnerEntity);
                     timeSlotEntity.setFieldTypeId(fieldTypeEntity);
-                    timeSlotEntity.setStartTime(timeEnableEntity.getStartTime());
-                    timeSlotEntity.setEndTime(timeEnableEntity.getEndTime());
-                    timeSlotEntity.setPrice(timeEnableEntity.getPrice());
+                    // Lấy startTime của timeEnable nhỏ nhất
+                    timeSlotEntity.setStartTime(timeEnableEntityList.get(0).getStartTime());
+                    // Lấy endTime của timeEnable lớn nhất
+                    timeSlotEntity.setEndTime(timeEnableEntityList.get(timeEnableEntityList.size() - 1).getEndTime());
                     timeSlotEntity.setReserveStatus(false);
                     timeSlotEntity.setStatus(true);
-                    // chủ sân có bao nhiêu sân thì tạo ra bấy nhiêu timeSlot tại 1 thời điểm
+                    // chủ sân có bao nhiêu sân thì tạo ra bấy nhiêu timeSlot
                     for (int i = 0; i < numberOfField; i++) {
-                        savedTimeSlotEntityList.add(timeSlotRepository.save(timeSlotEntity));
+                        if (i == 0) {
+                            savedTimeSlotEntityList.add(timeSlotRepository.save(timeSlotEntity));
+                        } else{
+                            TimeSlotEntity moreTimeSlotEntity = new TimeSlotEntity();
+                            moreTimeSlotEntity.setDate(date);
+                            moreTimeSlotEntity.setFieldOwnerId(fieldOwnerEntity);
+                            moreTimeSlotEntity.setFieldTypeId(fieldTypeEntity);
+                            moreTimeSlotEntity.setStartTime(timeSlotEntity.getStartTime());
+                            moreTimeSlotEntity.setEndTime(timeSlotEntity.getEndTime());
+                            moreTimeSlotEntity.setReserveStatus(false);
+                            moreTimeSlotEntity.setStatus(true);
+                            savedTimeSlotEntityList.add(timeSlotRepository.save(moreTimeSlotEntity));
+                        }
                     }
                 }
             }
@@ -88,7 +99,7 @@ public class TimeSlotServices {
         }
         Date startTime = DateTimeUtils.convertFromStringToTime(inputReservationDTO.getStartTime());
 
-        Date endTime = new Date(startTime.getTime() + inputReservationDTO.getDuration()*60000);
+        Date endTime = new Date(startTime.getTime() + inputReservationDTO.getDuration() * 60000);
 //        endTime.setTime(startTime.getTime() + inputReservationDTO.getDuration() * 60000);
         // get tất cả thời gian rảnh theo chủ sân, loại sân và ngày
         List<TimeSlotEntity> timeSlotEntityList = timeSlotRepository.findByFieldOwnerIdAndFieldTypeIdAndDateAndReserveStatusAndStatus(fieldOwner, fieldType, targetDate, false, true);
@@ -96,42 +107,48 @@ public class TimeSlotServices {
             boolean checkStart = timeSlotEntity.getStartTime().before(startTime) || timeSlotEntity.getStartTime().equals(startTime);
             boolean checkEnd = timeSlotEntity.getEndTime().after(endTime) || timeSlotEntity.getEndTime().equals(endTime);
             if (checkStart && checkEnd) {
+                float price = calculatePriceForTimeSlot(startTime, inputReservationDTO.getDuration(), targetDate, fieldOwner, fieldType);
                 if (timeSlotEntity.getStartTime().before(startTime) && timeSlotEntity.getEndTime().after(endTime)) {
                     // khoảng thời gian trước giờ đặt
                     TimeSlotEntity timeSlotEntity1 = new TimeSlotEntity(fieldOwner, fieldType, targetDate,
-                            timeSlotEntity.getStartTime(), startTime, timeSlotEntity.getPrice(), false, true);
+                            timeSlotEntity.getStartTime(), startTime, Float.valueOf(0), false, true);
                     // khoảng thời gian đặt
                     TimeSlotEntity timeSlotEntity2 = new TimeSlotEntity(fieldOwner, fieldType, targetDate,
-                            startTime, endTime, timeSlotEntity.getPrice(), true, true);
+                            startTime, endTime, price, true, true);
                     // khoảng thời gian sau giờ đặt
                     TimeSlotEntity timeSlotEntity3 = new TimeSlotEntity(fieldOwner, fieldType, targetDate,
-                            endTime, timeSlotEntity.getEndTime(), timeSlotEntity.getPrice(), false, true);
+                            endTime, timeSlotEntity.getEndTime(), Float.valueOf(0), false, true);
 
                     timeSlotRepository.delete(timeSlotEntity);
                     timeSlotRepository.save(timeSlotEntity1);
                     timeSlotRepository.save(timeSlotEntity3);
 
                     return friendlyMatchServices.createNewFriendlyMatch(timeSlotRepository.save(timeSlotEntity2), inputReservationDTO.getUserId());
-                } else if (timeSlotEntity.getStartTime().before(startTime)) {
+                } else if (timeSlotEntity.getStartTime().before(startTime) && timeSlotEntity.getEndTime().equals(startTime)) {
                     // khoảng thời gian trước giờ đặt
                     TimeSlotEntity timeSlotEntity1 = new TimeSlotEntity(fieldOwner, fieldType, targetDate,
-                            timeSlotEntity.getStartTime(), startTime, timeSlotEntity.getPrice(), false, true);
+                            timeSlotEntity.getStartTime(), startTime, Float.valueOf(0), false, true);
                     // khoảng thời gian đặt
                     TimeSlotEntity timeSlotEntity2 = new TimeSlotEntity(fieldOwner, fieldType, targetDate,
-                            startTime, endTime, timeSlotEntity.getPrice(), true, true);
+                            startTime, endTime, price, true, true);
                     timeSlotRepository.delete(timeSlotEntity);
                     timeSlotRepository.save(timeSlotEntity1);
                     return friendlyMatchServices.createNewFriendlyMatch(timeSlotRepository.save(timeSlotEntity2), inputReservationDTO.getUserId());
-                } else {
+                } else if (timeSlotEntity.getStartTime().equals(startTime) && timeSlotEntity.getEndTime().after(endTime)) {
                     // khoảng thời gian đặt
                     TimeSlotEntity timeSlotEntity1 = new TimeSlotEntity(fieldOwner, fieldType, targetDate,
-                            startTime, endTime, timeSlotEntity.getPrice(), true, true);
+                            startTime, endTime, price, true, true);
                     // khoảng thời gian sau giờ đặt
                     TimeSlotEntity timeSlotEntity2 = new TimeSlotEntity(fieldOwner, fieldType, targetDate,
-                            endTime, timeSlotEntity.getEndTime(), timeSlotEntity.getPrice(), false, true);
+                            endTime, timeSlotEntity.getEndTime(), Float.valueOf(0), false, true);
                     timeSlotRepository.delete(timeSlotEntity);
                     timeSlotRepository.save(timeSlotEntity2);
                     return friendlyMatchServices.createNewFriendlyMatch(timeSlotRepository.save(timeSlotEntity1), inputReservationDTO.getUserId());
+                } else {
+                    // khoảng thời gian đặt bằng chính timeslot
+                    timeSlotEntity.setReserveStatus(true);
+                    timeSlotEntity.setPrice(price);
+                    return friendlyMatchServices.createNewFriendlyMatch(timeSlotRepository.save(timeSlotEntity), inputReservationDTO.getUserId());
                 }
             }
         }
@@ -143,7 +160,14 @@ public class TimeSlotServices {
         FieldTypeEntity fieldType = fieldTypeServices.findById(fieldTypeId);
         Date targetDate = DateTimeUtils.convertFromStringToDate(dateString);
         if (timeSlotRepository.countByFieldOwnerIdAndFieldTypeIdAndDateAndStatus(fieldOwner, fieldType, targetDate, true) == 0) {
-            return createTimeSlotForDate(targetDate);
+            List<TimeSlotEntity> timeSlotEntityListAllFieldType = createTimeSlotForDate(targetDate);
+            List<TimeSlotEntity> timeSlotEntityList = new ArrayList<>();
+            for (TimeSlotEntity timeSlotEntity : timeSlotEntityListAllFieldType) {
+                if (timeSlotEntity.getFieldTypeId().getId() == fieldTypeId) {
+                    timeSlotEntityList.add(timeSlotEntity);
+                }
+            }
+            return timeSlotEntityList;
         } else {
             List<TimeSlotEntity> timeSlotEntityList = timeSlotRepository.findByFieldOwnerIdAndFieldTypeIdAndDateAndReserveStatusAndStatus(fieldOwner,
                     fieldType, targetDate, false, true);
@@ -151,21 +175,71 @@ public class TimeSlotServices {
         }
     }
 
+    private Float calculatePriceForTimeSlot(Date startTime, int duration, Date targetDate, AccountEntity fieldOwner, FieldTypeEntity fieldType) {
+        // chia thời gian thành những khoảng nhỏ 30 phút
+        int numberOfTimeSlide = duration / 30;
+        List<TimeSlotDTO> timeSlotDTOList = new ArrayList<>();
+        for (int i = 0; i < numberOfTimeSlide; i++) {
+            TimeSlotDTO timeSlotDTO = new TimeSlotDTO();
+            if (i == 0) {
+                timeSlotDTO.setStartTime(startTime);
+            } else {
+                timeSlotDTO.setStartTime(timeSlotDTOList.get(i - 1).getEndTime());
+            }
+            timeSlotDTO.setEndTime(new Date(timeSlotDTO.getStartTime().getTime() + 30 * 60000));
+            timeSlotDTOList.add(timeSlotDTO);
+        }
 
-//    public TimeSlotEntity reserveFriendlyMatch(InputReservationDTO inputFriendlyMatch){
-//        AccountEntity fieldOwnerEntity = accountRepository.findByIdAndStatus(inputFriendlyMatch.getFieldOwnerId(), true);
-//        FieldTypeEntity fieldTypeEntity = fieldTypeServices.findFieldTypeEntityById(inputFriendlyMatch.getFieldTypeId());
-//        List<FieldEntity> fieldEntityList = fieldRepository.findByFieldOwnerIdAndFieldTypeIdAndStatus(fieldOwnerEntity, fieldTypeEntity, true);
-//        // tìm list timeslot phù hợp
-//        List<TimeSlotEntity> allTimeSlotEntity = new ArrayList<>();
-//        for (FieldEntity fieldEntity : fieldEntityList) {
-//            List<TimeSlotEntity> timeSlotEntityList = timeSlotRepository.findByDateAndFieldIdAndReserveStatusAndStatus(inputFriendlyMatch.getDate(),
-//                    fieldEntity, false, true);
-//            allTimeSlotEntity.addAll(timeSlotEntityList);
-//        }
-//        for (TimeSlotEntity timeSlotEntity : allTimeSlotEntity) {
-//
-//        }
-//    }
+        String dateInWeek = DateTimeUtils.returnDayInWeek(targetDate);
+
+        List<TimeEnableEntity> timeEnableEntityList = timeEnableServices.findTimeEnableByFieldOwnerTypeAndDate(fieldOwner, fieldType, dateInWeek);
+        Float priceReturn = Float.valueOf(0);
+        for (TimeSlotDTO timeSlotDTO : timeSlotDTOList) {
+            for (TimeEnableEntity timeEnableEntity : timeEnableEntityList) {
+                boolean checkStart = timeEnableEntity.getStartTime().before(timeSlotDTO.getStartTime()) || timeEnableEntity.getStartTime().equals(timeSlotDTO.getStartTime());
+                boolean checkEnd = timeEnableEntity.getEndTime().after(timeSlotDTO.getEndTime()) || timeEnableEntity.getEndTime().equals(timeSlotDTO.getEndTime());
+                if (checkStart && checkEnd) {
+                    timeSlotDTO.setPrice(timeEnableEntity.getPrice() / 2);
+                }
+            }
+            priceReturn += timeSlotDTO.getPrice();
+        }
+        return priceReturn;
+    }
+
+    public List<FieldEntity> getListFreeFieldAtSpecificTime(String targetDateStr, String targetTimeStr, int fieldOwnerId, int fieldTypeId){
+        AccountEntity fieldOwner = accountServices.findAccountEntityById(fieldOwnerId, "owner");
+        FieldTypeEntity fieldType = fieldTypeServices.findById(fieldTypeId);
+        Date targetDate = DateTimeUtils.convertFromStringToDate(targetDateStr);
+        Date targetTime = DateTimeUtils.convertFromStringToTime(targetTimeStr);
+
+        List<FieldEntity> fieldEntityList = fieldServices.findFieldEntityByFieldOwnerAndFieldType(fieldOwner, fieldType);
+        List<TimeSlotEntity> timeSlotEntityList = timeSlotRepository.findTimeSlotHaveMatch(fieldOwner, fieldType, targetDate, targetTime, true, true);
+        List<FieldEntity> returnFieldEntityList = new ArrayList<>();
+        for (FieldEntity fieldEntity : fieldEntityList) {
+            boolean check = true;
+            for (TimeSlotEntity timeSlotEntity : timeSlotEntityList) {
+                if(timeSlotEntity.getFieldId()!= null && timeSlotEntity.getFieldId().getId() == fieldEntity.getId()){
+                    check = false;
+                }
+            }
+            if(check){
+               returnFieldEntityList.add(fieldEntity);
+            }
+        }
+        return returnFieldEntityList;
+    }
+
+    public TimeSlotEntity findById(int timeSlotId){
+        return timeSlotRepository.findByIdAndStatus(timeSlotId, true);
+    }
+
+    public TimeSlotEntity setTimeForTimeSlot(int timeSlotId, int fieldId){
+        TimeSlotEntity timeSlotEntity = findById(timeSlotId);
+        FieldEntity fieldEntity = fieldServices.findFieldEntityById(fieldId);
+        timeSlotEntity.setFieldId(fieldEntity);
+        return timeSlotRepository.save(timeSlotEntity);
+    }
+
 
 }
